requirements.txt
opencv-python
mediapipe
customtkinter
Pillow
numpy